add cryptocurrencyhistory functionality
- on manual price change of crypto, create history
- cryptocurrencychistory controller endpoints
- added connections and field defaults to dbcontext
- changed price and amount values to decimals to better reflect real life transactions in all entities
- added dtos for cryptocurrencyhistory
- cryptocurrency entity fields data annotated
- cryptocurrencyhistory entity fields data annotated
- user entity fields data annotated
- cryptocurrencyhistory profile added functionality
- cryptocurrency profile revised, added values for fields on creation

+++ CryptoService/DTOs/CryptocurrencyDTO.cs
@@ -6,9 +6,9 @@ public class CryptocurrencyGetDTO
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public float StartingPrice { get; set; }
-    public float CurrentPrice { get; set; }
-    public int TotalAmount { get; set; }
+    public decimal StartingPrice { get; set; }
+    public decimal CurrentPrice { get; set; }
+    public decimal TotalAmount { get; set; }
 }
 
 public class CryptocurrencyGetByIdDTO
@@ -16,10 +16,10 @@ public class CryptocurrencyGetByIdDTO
     public int Id { get; set; }
     public string Name { get; set; }
     public string ShortName { get; set; }
-    public float StartingPrice { get; set; }
-    public float CurrentPrice { get; set; }
-    public int TotalAmount { get; set; }
-    public int AvailableAmount { get; set; }
+    public decimal StartingPrice { get; set; }
+    public decimal CurrentPrice { get; set; }
+    public decimal TotalAmount { get; set; }
+    public decimal AvailableAmount { get; set; }
     public List<CryptocurrencyHistory> CryptocurrencyHistory { get; set; }
 }
 
@@ -27,14 +27,14 @@ public class CryptocurrencyCreateDTO
 {
     public string Name { get; set; }
     public string? ShortName { get; set; }
-    public float StartingPrice { get; set; }
-    public int TotalAmount { get; set; }
+    public decimal StartingPrice { get; set; }
+    public decimal TotalAmount { get; set; }
 }
 
 public class CryptocurrencyUpdateDTO
 {
     public string? Name { get; set; }
     public string? ShortName { get; set; }
-    public float? CurrentPrice { get; set; }
-    public int? TotalAmount { get; set; }
+    public decimal? CurrentPrice { get; set; }
+    public decimal? TotalAmount { get; set; }
 }
